feat: Seeder for major - subject

+++ app/Http/Controllers/Academics/MajorController.php
@@ -87,6 +87,12 @@ class MajorController extends Controller
    */
   public function destroy(Major $major)
   {
+    if ($major->students()->count() > 0) :
+      return response()->json([
+        'message' => trans('session.delete_error')
+      ], 400);
+    endif;
+
     $this->majorService->delete($major->id);
     return response()->json([
       'message' => trans('session.delete'),
